build: hold every directory to -Wall -Wextra -Werror unless it declares the sites it still carries
share/mk/warnings.mk selects WARNERR from WARNLEVEL: full, the default,
is -Wall -Wextra -Werror, and legacy is -Werror alone. A leaf Makefile
assigns WARNLEVEL=legacy before its sys.mk or tools/Makefile.inc
include, because both compose CC with != at include time and a later
assignment selects nothing; check-warning-policy-host now reads every
tracked Makefile and fails an assignment that follows its include or
names another level. tools/warning-census.sh overrides WARNERR on the
command line, which wins over both levels, so the census still measures
every directory at one level.

The census on the repaired headers is 4722 instances at 4348 sites over
1394 cross compiles. 99 of the 241 leaf directories it compiles raise
nothing under both groups and build at the full level; 142 declare
legacy with the count of distinct sites the census measured there in
the comment above the assignment, so git grep WARNLEVEL is the ledger of
what remains. Six tool directories and usr.bin/as/tests, outside the
census, were measured by hand with the same override and declare the
same way. share/zoneinfo repairs two signed comparisons in zdump.c and
marks zic.c's usage() noreturn; romprobe hides the constant ROM address
from the array-bounds folder through an empty asm; sigtest casts its
three-argument handler through void *.

Two more ways the census misled are recorded in the document: the
classifier read a compile that links in the same command, as
usr.bin/retroforth does, as a link and counted its warnings nowhere, so
a cross compile is now any cross line carrying -c or naming a .c or .S
source, calibrated on ten lines under sh and dash; and the 48
builtin-mismatch sites attributed to three tree headers were cron's
compat.h removing const before the headers were read.

check-warning-policy-cross compiles the -Wall and -Wextra probes on the
full routes bin/echo, usr.bin/smlrc and lib/libc, where both are fatal,
and on the legacy routes bin/sh and usr.bin/uucp, where the -Wextra
probe must still produce an object. Verified from clean: bmake
MACHINE=rp2040 distribution, check-board-build, check-host and
check-cross all pass.

+++ share/zoneinfo/zdump.c
@@ -20,14 +20,14 @@ extern char *tzname[2];
 extern char *imalloc(int);
 extern void tzset(void);
 
-static int	longest;
+static size_t	longest;
 
 static void
 show(char *zone, time_t t, int v)
 {
 	struct tm *		tmp;
 
-	(void) printf("%-*s  ", longest, zone);
+	(void) printf("%-*s  ", (int) longest, zone);
 	if (v)
 		(void) printf("%.24s GMT = ", asctime(gmtime(&t)));
 	tmp = localtime(&t);
@@ -126,7 +126,7 @@ main(int argc, char *argv[])
 			fp = fopen(argv[i], "r");
 		else {
 			j = strlen(TZDIR) + 1 + strlen(argv[i]) + 1;
-			if (j > sizeof buf) {
+			if ((size_t) j > sizeof buf) {
 				(void) fprintf(stderr,
 					"%s: timezone name %s/%s is too long\n",
 					argv[0], TZDIR, argv[i]);
